rename package rtkutils -> gpsutils
