+++ components/movementsensor/gpsutils/cachedData.go
@@ -1,4 +1,4 @@
-package rtkutils
+package gpsutils
 
 import (
 	"context"
