+++ components/movementsensor/gpsutils/cachedData_test.go
@@ -1,4 +1,4 @@
-package rtkutils
+package gpsutils
 
 import (
 	"context"
