+++ components/movementsensor/gpsutils/nmeaParser.go
@@ -1,4 +1,4 @@
-package rtkutils
+package gpsutils
 
 import (
 	"fmt"
